Playing with beat timings
Map beat configs' time should be configured so that the given time
is the time when the beat hits the target (not when it spawns).
We need to carefully animate the falling beats, so that we
know when it will hit the target, so that we know which
time offset we can use when spawning the beats.

Add dev map with some beat configs.